update the api surface
