+++ Assets/MixedRealityToolkit.SDK/Experimental/Features/Audio/Scripts/WindowsMicrophoneStreamErrorCodes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3893c32a24c4ee44993ea33aea707fb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
